feat: add validation to the fields of edit-address-form, rename this file
